+++ include/StrongAI/World/DarwinTicTacToe/TicTacToe.hpp
@@ -13,10 +13,10 @@ class TicTacToe
 {
     public:
         enum class Token { None, X, O };
-        const float ILLEGAL_MOVE_WORTH = -1.f;
-        const float LOST_GAME_WORTH = -.5f;
-        const float TIE_GAME_WORTH = 0.f;
-        const float WON_GAME_WORTH = 1.f;
+        const double ILLEGAL_MOVE_WORTH = -1.f;
+        const double LOST_GAME_WORTH = -.5f;
+        const double TIE_GAME_WORTH = 0.f;
+        const double WON_GAME_WORTH = 1.f;
 
         TicTacToe() {}
         virtual ~TicTacToe() {}
@@ -30,12 +30,12 @@ class TicTacToe
         /**
         *   Converts the player output to an integer in range [ 0, 8 ]
         **/
-        int getPlayerMove( GeneralAI& player, const std::vector< float >& input );
+        int getPlayerMove( GeneralAI& player, const std::vector< double >& input );
 
         ///End game conditions
         bool isWinning( const std::vector< Token >& board, Token turn );
         bool boardIsFull( const std::vector< Token >& board );
-        std::vector< float > boardToInput( const std::vector< Token >& board );
+        std::vector< double > boardToInput( const std::vector< Token >& board );
 };
 
 #endif // TIC_TAC_TOE_HPP
