+++ include/StrongAI/World/TicTacToe/TicTacToeHuman.hpp
@@ -12,11 +12,11 @@ class TicTacToeHuman : public HumanAI
         TicTacToeHuman( int inputSize, int outputSize, int maxInput, int maxOutput );
 
     private:
-        std::vector< float > coreOutput( const std::vector< float >& input );
-        void coreLearn( const std::vector< float >& input, const std::vector< float >& output, float outcome ) {}
+        std::vector< double > coreOutput( const std::vector< double >& input );
+        void coreLearn( const std::vector< double >& input, const std::vector< double >& output, double outcome ) {}
 
-        void visualizeInput( const std::vector< float >& input );
-        char inputToChar( float input );
+        void visualizeInput( const std::vector< double >& input );
+        char inputToChar( double input );
 };
 
 #endif // TIC_TAC_TOE_HUMAN_HPP
